==> script/1. cleaningdata.rmd <==
---
title: "cleaning data"
author: "Thomas Delcey"
date: '2022-06-27'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

-   [1 What is this script for?](#1 What-is-this-script-for)
-   [2 Loading packages, paths and data](#2 Loading-packages-paths-and-data)
-   [3 Cleaning data](#3 Cleaning data)


#1 What is this script for?


#2 Loading packages, paths and
    data
    
```{r, echo=FALSE results = "hide"}

source("C:/Users/thomd/Documents/MEGA/github/ANR/script/0_paths_and_packages.R")

```

```{r, results = "hide"}

file.list <- paste0(data_path, list.files(path = data_path, pattern='*.xlsx'))
df.list <- lapply(file.list, read_excel)

dgpie_partenaire_df <- df.list[[1]]
dgpie_projet_df <- df.list[[2]]
dos_partenaire_df <- df.list[[3]]
dos_projet_df <- df.list[[4]]
discipline_code_df <- df.list[[5]]

```

#3 Cleaning_data 

## renaming variables

```{r}
dos_projet_df <- dos_projet_df %>%
  rename(text = Projet.Resume.Anglais, 
         year = AAP.Edition,
         id = Projet.Code_Decision_ANR,
         ) 

dos_partenaire_df <- dos_partenaire_df %>%
  rename(name = Projet.Partenaire.Responsable_scientifique.Prenom,
         lastname = Projet.Partenaire.Responsable_scientifique.Nom) 
```


##3.1 Filter general call & create a discipline variable 

```{r echo=FALSE results = "hide"}
dos_projet_df <- dos_projet_df %>%
  mutate(CES = str_extract(id, "CE[:digit:]{2}")) %>% left_join(discipline_code_df, by = 'CES') %>%
  mutate(generique = case_when(str_detect(Programme.Acronyme, "^AAPG") ~ 1, 
                               TRUE ~ 0
                              )
        ) %>%
  mutate(social_science = case_when(str_detect(CES, "CE(26|27|28|41)") ~ 1, 
                               TRUE ~ 0
                              )
        )

saveRDS(dos_projet_df, paste0(data_path, "dos_projet_df.RDS"))

```



## 3.2 Institution 

'Institution' needs to clean. There is first of all acronyms and typo mistakes. Second, we need to aggregate local institutions such as 'CNRS Paris' and 'CNRS Lyon'

We first want to apply basic cleaning such as removing punctions, upper case and so on.

```{r echo=FALSE}

#UNIFINISH 

dos_partenaire_df <- dos_partenaire_df %>% 
  right_join(dos_projet_df) %>% #keep only general call
  filter(Projet.Partenaire.Est_coordinateur == 'TRUE') %>% #keep only project leader 
  mutate(institution = tolower(Projet.Partenaire.Nom_organisme)) %>% #rename and remove uppercase 
  mutate(institution = str_remove(institution, "[:punct:]")) %>% #remove punction 
  mutate(institution = case_when(str_detect(institution, "^aix(-| |)marseille") ~ "Aix Marseille",
                                 #str_detect(institution, "")
                                 #str_detect(institution, "") ~ "",
                                 TRUE ~ dos_partenaire_df$institution)
         ) %>% select(institution) 

```

Calculate distance between character vectors and merge lines that matchh 

```{r echo=FALSE}
df_save <- dos_partenaire_df[0,]
for (i in (1:nrow(dos_partenaire_df))){
   df_match <- dos_partenaire_df[agrep(dos_partenaire_df$institution[[i]], dos_partenaire_df$institution, max.distance = 0.05),]
   df_match <- df_match %>% mutate(institution = institution[[1]])
   df_save <- rbind(df_save, df_match)
    df_save <- df_save[!duplicated(df_save),]
    print(i)
}
```

##3.3 create a gender variable 

```{r, echo=FALSE}
library("gender")
library("genderdata")


##cleaning the prenom variable

#loading the CLEAN data 

dos_partenaire_df <- readRDS(paste0(data_path, "dos_partenaire_df.RDS"))


##function of Etienne for saving 
Unaccent <- function(text) {
  text <- gsub("['`^~\"]", " ", text)
  text <- iconv(text, to="ASCII//TRANSLIT//IGNORE")
  text <- gsub("['`^~\"]", "", text)
  return(text)
}

##removing accent 
dos_partenaire_df<- dos_partenaire_df %>%
  mutate(name = unaccent(name))

#keep only the first_name
dos_partenaire_df <- dos_partenaire_df %>%
  mutate(name = str_extract(name, "[a-zA-Z]+"))


#detect gender
  
name <- as.character(dos_partenaire_df$name)
gender_df <- gender(name) 

dos_partenaire_df <- dos_partenaire_df %>%
  left_join(select(gender_df, c(name, gender))) %>%
  distinct()

dos_partenaire_df %>%
  group_by(gender) %>%
  summarise(n = n()) %>%
  mutate('%' = n/sum(n)*100)


saveRDS(dos_partenaire_df, paste0(data_path, "dos_partenaire_df.RDS")) 


dos_projet_df <- readRDS(paste0(data_path, "dos_projet_df.RDS"))

dos_projet_df_g <- dos_projet_df %>%
  select(year, id, Projet.Aide_allouee, CES, Category) %>%
  left_join(select(dos_partenaire_df, c(id, gender))) %>%
  filter(!is.na(Category)) %>%
  filter(!is.na(gender))

gender_by_category <- dos_projet_df_g %>%
  count(year, gender, Category)

gender_by_CES <- dos_projet_df_g %>%
  count(year, gender, CES)

gender_by_year <- dos_projet_df_g %>%
  count(year, gender)

gender_by_grant <- dos_projet_df_g %>%
  group_by(year, gender) %>%
  summarise(grant = sum(Projet.Aide_allouee))

saveRDS(gender_by_category, paste0(data_path, "gender_by_category.RDS"))
saveRDS(gender_by_CES, paste0(data_path, "gender_by_CES.RDS"))
saveRDS(gender_by_year, paste0(data_path, "gender_by_year.RDS"))
saveRDS(gender_by_grant, paste0(data_path, "gender_by_grant.RDS"))

```
